mvoe to_api_duration into ts
